# sales-portal-order/docs/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "orders",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://devapi-pegasus.tataconsumer.com",
      "variables": {}
    },
    {
      "url": "https://uatapi-pegasus.tataconsumer.com",
      "variables": {}
    },
    {
      "url": "https://prdapi-pegasus.tataconsumer.com",
      "variables": {}
    }
  ],
  "paths": {
...
    "/order/api/v1/util/update-material-tags": {
      "post": {
        "tags": ["Misc"],
        "summary": "This api is used to update the material tags provided in a excel from the request body in the product/material master data in the database in order to enhance the search operation",
        "operationId": "update-material-tags",
        "consumes": "multipart/form-data",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/updatematerialtagsrequest",
                "example": {
                  "dataset": "/home/Downloads/Ordering Portal Material List with tagging_Updated_dt 26.11.2021.xlsx"
                },
                "properties": {
                  "media": {
                    "type": "string",
                    "format": "base64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "updatematerialtagsrequest": {
        "title": "updatematerialtagsrequest",
        "required": ["dataset"],
        "type": "object",
        "properties": {
          "dataset": {
            "type": "file"
          }
        },
        "example": {
          "dataset": "/home/Downloads/Ordering Portal Material List with tagging_Updated_dt 26.11.2021.xlsx"
        }
      },
...
    }
  },
  "security": [{ "api_key": [] }],
...
}